Simplfy classification
