+++ classification/rego/dob.rego
@@ -7,39 +7,39 @@ output[k] := v if {
 	v := classify(k, input[k])
 }
 
-default classify(_, _) := "UNLABELED"
+default classify(_, _) := false
 
-classify(key, _) := "DOB" if {
+classify(key, _) if {
 	lower(key) == "dob"
 }
 
-classify(key, _) := "DOB" if {
+classify(key, _) if {
 	lower(key) == "dateofbirth"
 }
 
-classify(key, _) := "DOB" if {
+classify(key, _) if {
 	lower(key) == "date_of_birth"
 }
 
-classify(key, _) := "DOB" if {
+classify(key, _) if {
 	lower(key) == "birthdate"
 }
 
-classify(key, _) := "DOB" if {
+classify(key, _) if {
 	lower(key) == "birth_date"
 }
 
-classify(_, val) := "DOB" if {
+classify(_, val) if {
 	# mm/dd/yyyy mm-dd-yyyy mm.dd.yyyy
 	regex.match(`^(0?[1-9]|1[0-2])[\/\.-](0?[1-9]|[12]\d|3[01])[\/\.-](19|20)\d{2}$`, val)
 }
 
-classify(_, val) := "DOB" if {
+classify(_, val) if {
 	# dd/mm/yyyy
 	regex.match(`^(0?[1-9]|[12]\d|3[01])[\/\.-](0?[1-9]|1[0-2])[\/\.-](19|20)\d{2}$`, val)
 }
 
-classify(_, val) := "DOB" if {
+classify(_, val) if {
 	# yyyy/mm/dd
 	regex.match(`^(19|20)\d{2}[\/\.-](0?[1-9]|1[0-2])[\/\.-](0?[1-9]|[12]\d|3[01])$`, val)
 }
